move StoryExample
