+++ website/extension-examples/extension-node-formatting/basic.tsx
@@ -10,7 +10,7 @@ import CodeBlock from '@theme/CodeBlock';
 import BrowserOnly from '@docusaurus/BrowserOnly';
 import ComponentSource from '!!raw-loader!../../../packages/storybook-react/stories/extension-node-formatting/basic.tsx';
 
-import { StoryExample } from '../../components/example-root';
+import { StoryExample } from '../../src/components/story-example-component';
 
 const ExampleComponent = (): JSX.Element => {
   const story = (
@@ -27,3 +27,4 @@ const ExampleComponent = (): JSX.Element => {
 };
 
 export default ExampleComponent;
+
