+++ website/src/components/story-example-component.tsx
